Infrastructure: move methods' opening brace to start of next line (#8465)
#### Brief overview of PR changes/additions
Some recently added contributions have not followed our coding style of
placing the open brace of a method/function in column 1 of the next line
after the rest of the text. This PR converts them to match our preferred
style which is not one that our `.clang-format` style definition can
produce.

#### Motivation for adding to Mudlet
Make our code more uniform in style.

#### Other info (issues closed, discussion etc)

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>