Infrastructure: refactor build to use static library on Linux/macOS (#8461)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Currently, the project links the source files directly to the
executable.
This PR refactors the build to compile the sources into a static library
(libmudlet_core.a on Linux), which is then linked to the executable.

#### Motivation for adding to Mudlet

1. This change lays the groundwork for a future split between libmudlet
and a front-end library for the mobile version #8030.
2. Simplifies the build setup for future functional tests
https://github.com/NicolasKeita/Mudlet/pull/1.
Rather than manually specifying source files and repeating the build
configuration in each test target, tests can now simply link against the
static library.

---------

Co-authored-by: Vadim Peretokin <vperetokin@hey.com>