Add: Happy-eyeballs (try IPv4 and IPv6 connections simultaneously) (#8135)
#### Brief overview of PR changes/additions
Overhauls the profile connection process to identify whether IPv4 and/or
IPv6 protocol addresses are available for the entered details and if
BOTH are to then try to connect with both at the same time (with a
slight preference for IPv6 {by starting it first}). The first one to
connect (including securely and/or via a proxy) wins out and terminates
the process using the other address protocol. This technique is known as
[Happy Eyeballs](https://en.wikipedia.org/wiki/Happy_Eyeballs).

#### Motivation for adding to Mudlet
To improve the end-user's experience if they are fortunate to be on a
"dual-stack" (both protocols available) network.

#### Other info (issues closed, discussion etc)
The on-screen messages have been massaged somewhat to inform the user
about the nature of the connection(s) being made. ALL the addresses that
have been found for the Server are listed instead of just the first one
{this includes a "numerus" translatable string which will need
processing by yours truly!} Should a raw IP address be provided instead
of a "URL" then the result of the "reverse" lookup (trying to get a host
name from the IP address) is displayed instead of just repeating the
address - if anything *is* found - if not that is also noted instead.

One other effect of Mudlet being given a raw IP address is that those
are not normally included in an SSL/TLS certificate - so rendering a
secure connection impossible in that case - this was already noted for
an IPv4 address but NOT for an IPv6 one. This PR corrects that to also
reject secure connections being attempted if an IPv6 address is provided
as the URL.

---------

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>