Fix: Copy from console now works when command line has focus (#8551)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions

When you select text in the main game window and press Ctrl+C (or Cmd+C
on Mac), that text is now copied to the clipboard even if the command
line has text or focus. Previously, the command line would take priority
and prevent copying from the main window unless the command line was
empty.

#### Motivation for adding to Mudlet

This has been a long-standing usability issue affecting many users. It
made copying text from the game window unnecessarily difficult, as users
had to clear the command line first or use the right-click menu instead
of the standard keyboard shortcut.

#### Other info (issues closed, discussion etc)

Closes #7114
See also: #3891, #7008, #6758, #6505, #7147, #7124, #7070, #4956, #4540,
#3922, #604, #6504