Fix: Windows updater file lock issue in dblsqd (prevented updates) (#8661)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Update dblsqd to a newer version that releases the lock on the file,
allowing the newly downloaded installer to be moved to the temporary
directory.
#### Motivation for adding to Mudlet
Part of the fixes in getting Windows updates working seamlessly.
#### Other info (issues closed, discussion etc)

(cherry picked from commit 208dc2b27a049620b8c452a400cfd7e774d03011)