Infrastructure: allow CMake `initGitSubmodule.cmake` to act recursively (#8717)
#### Brief overview of PR changes/additions
Add a `RECURSIVE` option to this CMake function (`initGitSubmodule()`)
to recursively clone all the submodules of a submodule.

#### Motivation for adding to Mudlet
To allow CMake builds of Mudlet to pull in a submodule that itself
contains its own submodules. This is intended to aid a revision to the
meta-build system so that end-users will not have to clone every git
submodule themselves but instead to restore the original design that
would do it. This also means it would no longer be needed to use the
`submodules: recursive` option in the GH `actions/checkout@v6` Action -
thereby speeding things up and reducing the disk space used a bit.

#### Other info (issues closed, discussion etc)
I need this to fixup developer builds - particular on Windows - where
and when the Sentry crash-reporting functionality is NOT required given
that it uses a number of submodules and subsubmodules of its own.

As presented in this PR the new option is not actually used so should
not alter the build in any way (other than to also add the "normal" text
output from running any `git submodule update --init` action which were
previously discarded0 but it is a section of code that can be done as a
separate item.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>